+++ zemfrog/generator.py
@@ -2,13 +2,8 @@ import os
 from flask.globals import current_app
 from jinja2 import Template
 
-from .helper import (
-    copy_template,
-    get_import_name,
-    get_template,
-    search_model,
-    validate_name,
-)
+from .helper import copy_template, get_import_name, get_template, search_model
+from .validators import validate_module_name
 
 
 def g_project(name: str, import_name: str):
@@ -19,7 +14,7 @@ def g_project(name: str, import_name: str):
 
     """
 
-    validate_name(name)
+    validate_module_name(name)
     print("Creating %r project... " % name, end="")
     copy_template("project", name)
     main_app = True if import_name == "wsgi" else False
@@ -40,6 +35,83 @@ def g_project(name: str, import_name: str):
     print("(done)")
 
 
+def g_extension(name: str):
+    """
+    Function for making extension boilerplate.
+
+    :param name: extension name.
+    """
+
+    validate_module_name(name)
+    print("Creating a %r extension..." % name, end="")
+    ext_dir = os.path.join(current_app.root_path, "extensions")
+    old_filename = get_template("extension", "name.py")
+    with open(old_filename) as fp:
+        new_data = fp.read()
+
+    new_filename = os.path.join(ext_dir, name.lower() + ".py")
+    with open(new_filename, "w") as fp:
+        fp.write(new_data)
+
+    print("(done)")
+
+
+def g_task(name: str):
+    """
+    Function for creating celery task.
+
+    :param name: task name.
+    """
+
+    validate_module_name(name)
+    print("Creating %r task..." % name, end="")
+    import_name = get_import_name(current_app)
+    main_app = True
+    if import_name:
+        main_app = False
+
+    ext_dir = os.path.join(current_app.root_path, "tasks")
+    old_filename = get_template("task", "name.py")
+    with open(old_filename) as fp:
+        old_data = fp.read()
+        py_t = Template(old_data)
+        new_data = py_t.render(main_app=main_app, task_name=name)
+
+    new_filename = os.path.join(ext_dir, name.lower() + ".py")
+    with open(new_filename, "w") as fp:
+        fp.write(new_data)
+
+    print("(done)")
+
+
+def g_model(name: str):
+    """
+    Function for creating sqlalchemy model.
+
+    :param name: model name.
+    """
+
+    validate_module_name(name)
+    print("Creating %r model..." % name, end="")
+    import_name = get_import_name(current_app)
+    main_app = True
+    if import_name:
+        main_app = False
+
+    model_dir = os.path.join(current_app.root_path, "models")
+    old_filename = get_template("model", "name.py")
+    with open(old_filename) as fp:
+        old_data = fp.read()
+        py_t = Template(old_data)
+        new_data = py_t.render(main_app=main_app, model_name=name.capitalize())
+
+    new_filename = os.path.join(model_dir, name.lower() + ".py")
+    with open(new_filename, "w") as fp:
+        fp.write(new_data)
+
+    print("(done)")
+
+
 def g_api(name: str):
     """
     Functions for creating APIs.
@@ -48,7 +120,7 @@ def g_api(name: str):
 
     """
 
-    validate_name(name)
+    validate_module_name(name)
     print("Creating API %r... " % name, end="")
     api_dir = os.path.join(current_app.root_path, "api")
     old_filename = get_template("api", "name.py")
@@ -72,7 +144,7 @@ def g_api_crud(name: str):
 
     """
 
-    validate_name(name)
+    validate_module_name(name)
     src_model = search_model(name)
     import_name = get_import_name(current_app)
     main_app = True
@@ -109,7 +181,7 @@ def g_blueprint(name: str):
 
     """
 
-    validate_name(name)
+    validate_module_name(name)
     print("Creating blueprint %r... " % name, end="")
     bp_dir = os.path.join(current_app.root_path, name.lower())
     copy_template("blueprint", bp_dir)
@@ -134,7 +206,7 @@ def g_middleware(name: str):
 
     """
 
-    validate_name(name)
+    validate_module_name(name)
     print("Creating middleware %r... " % name, end="")
     middleware_dir = os.path.join(current_app.root_path, "middlewares")
     old_filename = get_template("middleware", "name.py")
@@ -156,7 +228,7 @@ def g_command(name: str):
 
     """
 
-    validate_name(name)
+    validate_module_name(name)
     print("Creating command %r..." % name, end="")
     cmd_dir = os.path.join(current_app.root_path, "commands")
     old_filename = get_template("command", "name.py")
@@ -180,7 +252,7 @@ def g_error_handler(name):
 
     """
 
-    validate_name(name)
+    validate_module_name(name)
     print("Creating error handler %r... " % name, end="")
     eh_dir = os.path.join(current_app.root_path, "handlers")
     old_filename = get_template("errorhandler", "name.py")