+++ zemfrog/templates/task/name.py
@@ -1,0 +1,5 @@
+from {{ "" if main_app else ".." }}extensions.celery import celery
+
+@celery.task
+def {{task_name}}():
+    pass