+++ HISTORY.rst
@@ -109,3 +109,12 @@ History
 * support multiple static files
 * Add an endpoint to validate the password reset token
 * fix `#37 <https://github.com/zemfrog/zemfrog/issues/37>`_
+
+
+1.2.5 (2020-11-18)
+------------------
+
+* NEW: add extension, model, task generator
+* Refactor Code: add model mixin
+* add command user, role & permission
+* FIX: auth logs